Started the Shooter Big Brother
Created code that stops shooter when a certain amount of balls pass through. Need to ask robin how to set the counter back to zero. Added these commands to a big brother command group.

"It was a bright cold day in April, and the clocks were striking thirteen." George Orwell, 1984.