Old untested changes
This entire branch can be deleted for now, but this branch adds in an equation that I would like to test with the practice bot. The equation is a linear equation that tells the shooter how far it needs to shoot based on how far away it is from the goal post.